Refactor flickity carousel component #31
